mock/requirements.txt
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
numpy
matplotlib
imageio
Pillow>=10.3.0
opencv-python
trimesh
scipy
gradio>=5.0.0
transformers
huggingface_hub
git+https://github.com/openai/CLIP.git